be/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
python-dotenv
pydantic
pyodbc
